Bài viết, bình luận anime ok

+++ DataModels/EF/Ratings.cs
@@ -8,7 +8,7 @@ namespace DataModels.EF
     {
         public int Id { get; set; }
         [Range(0, 10)]
-        public int RatePoint { get; set; }
+        public double RatePoint { get; set; }
         [ForeignKey(nameof(User))] public int UserId { get; set; }
         public Users User { get; set; }
         [ForeignKey(nameof(Anime))] public int AnimeId { get; set; }
